Eliminación de script AnomalyCameraManager
He puesto lo de este script en CameraManager pues no tenia sentido que estuviesen separados.
