--- CameraPrototype/Assets/_Scripts/Camera/Anomalies Camera/AnomalyCameraManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 880d3d5aefaece34ca5ad26ad1e43578
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
